Python script to plot random map coordinates
This set of points was created by selecting a region in Bloomington, Indiana in a public park where there are no private residences as far as we can tell. The purpose of these points is to help us create a regional heat map. It is not required that these points be a source of truth for incident data from real life; in fact it's better that they aren't! The file being committed is simply a small Python script created by ChatGPT. When given coordinates that define a rectangular geographic area, this program randomly creates points within that area. After generating the points, I had them added to OpenStreetMap, which showed they were indeed confined to the park I selected.